scope3-scraper/requirements.txt
azure-functions
azure-storage-blob
playwright
openai
